# perf(weave): add indexed datetime filtering to calls query