--- a/weave/trace_server/calls_query_builder/utils.py
+++ b/weave/trace_server/calls_query_builder/utils.py
@@ -1,0 +1,20 @@
+import logging
+
+import sqlparse
+
+
+def param_slot(param_name: str, param_type: str) -> str:
+    """Helper function to create a parameter slot for a clickhouse query."""
+    return f"{{{param_name}:{param_type}}}"
+
+
+def safely_format_sql(
+    sql: str,
+    logger: logging.Logger,
+) -> str:
+    """Safely format a SQL string with parameters."""
+    try:
+        return sqlparse.format(sql, reindent=True)
+    except:
+        logger.info(f"Failed to format SQL: {sql}")
+        return sql