--- a/weave/trace_server/calls_query_builder/utils.py
+++ b/weave/trace_server/calls_query_builder/utils.py
@@ -1,4 +1,6 @@
+import contextlib
 import logging
+from collections.abc import Generator
 
 import sqlparse
 
@@ -18,3 +20,33 @@ def safely_format_sql(
     except:
         logger.info(f"Failed to format SQL: {sql}")
         return sql
+
+
+# Context for tracking if we're in a NOT operation
+class NotContext:
+    # Nesting depth
+    _depth = 0
+
+    @classmethod
+    @contextlib.contextmanager
+    def not_context(cls) -> Generator[None, None, None]:
+        """Context manager for NOT operations.
+
+        Properly handles nested NOT operations by tracking depth.
+        In boolean logic:
+        - NOT(expr) flips the result
+        - NOT(NOT(expr)) is equivalent to expr
+        - NOT(NOT(NOT(expr))) is equivalent to NOT(expr)
+
+        So we only apply special handling when nesting depth is odd.
+        """
+        cls._depth += 1
+        try:
+            yield
+        finally:
+            cls._depth -= 1
+
+    @classmethod
+    def is_in_not_context(cls) -> bool:
+        """Check if we're in a NOT context with odd nesting depth."""
+        return cls._depth % 2 == 1