… tests/trace/test_weave_client.py
… tests/trace_server/test_calls_query_builder.py
… weave/trace_server/clickhouse_trace_server_batched.py
… weave/trace_server/migrations/012_add_sortable_timestamp.down.sql
… weave/trace_server/migrations/012_add_sortable_timestamp.up.sql