## Description

<!--
Include reference to internal ticket "Fixes WB-NNNNN" and/or GitHub issue "Fixes #NNNN" (if applicable)
-->

[WB-24385](https://wandb.atlassian.net/browse/WB-24385)

This pr: 
- adds a new column `sortable_datetime` to the `calls_merged` table with default `coalesce(started_at, ended_at, NOW())`
- adds that column to the calls_merged_view as `coalesce(started_at, ended_at, call_parts.created_at)`
- adds and materialized a minmax index on that column
- refactors the sql optimization paradigm in `calls_query_builder` to be more of a mapper pattern
- adds a `SortableDatetimeOptimizationProcessor` class that adds datetime filtering pre-where to the calls query
- add a `NotContext` to track if the current filter has a `not` prepended. We need to do this so we can apply the buffer in the correct direction. 
- uses configurable timestamp buffer, defaults to 5 minutes. This means that in our pre-groupby filter we include more calls than necessary, and rely on the outer filter in the group by to pair down the results to be exact. 

NOTE:
- technically using this column as a heuristic is actually incorrect because we will occasionally filter out call_ends that are outside of the time window while the call_start is valid and within bounds. This will result in wonky results on the margins of the date filter. One easy way to lessen the impact of this is to make the initial date filter less strict (by adding a fuzzing/buffer to the initial timestamp filter). Because this is being implemented as an optimization, we still retain the original filter. 

Example:
```sql
SELECT count() as c
FROM
(
    SELECT calls_merged.id AS id
    FROM calls_merged
    WHERE calls_merged.project_id = 'project_id'
         AND calls_merged.sortable_datetime > '2025-04-01 03:25:58.0000'  --<<--------
    GROUP BY (calls_merged.project_id, calls_merged.id)
    HAVING (any(calls_merged.started_at) > _CAST(1742472000, 'UInt64')) AND (NOT (any(calls_merged.started_at) > _CAST(1742475600, 'UInt64'))) AND (any(calls_merged.deleted_at) IS NULL) AND (NOT (any(calls_merged.started_at) IS NULL))
)
```


## Testing

Updates and adds unit and client end to end tests 

Replica tests (stacked on `op_name` improvement already merged [here](https://github.com/wandb/weave/pull/4116))
| Environment | Num Runs | Avg Mem     | Rows Read | p50 Latency | p90 Latency | p99 Latency |
|-------------|----------|-------------|-----------|-------------|-------------|-------------|
| prod        | 5        | 9.24 GiB    | 8,271,114 | 2.895s      | 5.212s      | 5.212s      |
| branch      | 6        | 195.52 MiB  | 104,851   | 0.629s      | 1.213s      | 1.213s      |


With a slightly better order (datetime first, then op_name) compared to current prod (no op_name optimization)
| Environment | Num Runs | Avg Mem     | Rows Read | p50 Latency | p90 Latency | p99 Latency |
|-------------|----------|-------------|-----------|-------------|-------------|-------------|
| prod        | 4        | 22.42 GiB   | 8,271,114 | 5.494s      | 5.597s      | 5.597s      |
| branch      | 5        | 183.06 MiB  | 104,851   | 0.549s      | 0.625s      | 0.625s      |


Migration tests in prod replica take `<100ms`

[WB-24385]: https://wandb.atlassian.net/browse/WB-24385?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ